+++ NEWS.md
@@ -1,6 +1,8 @@
 # smdocker 0.1.1.9000
 
 * add github url to description
+* check execute role when in space app
+* guess execution role when lack of IAM permissions
 
 # smdocker 0.1.1